+++ R/openmp-utils.R
@@ -1,0 +1,7 @@
+setthreads <- function(threads) {
+    .Call(CsetThreads, as.integer(threads[1L]))
+}
+
+getthreads <- function() {
+    .Call(CgetThreads)
+}